pyproject.toml
[tool.poetry]
name = "ExVo-Fewshot"
version = "0.1.0"
description = ""
authors = ["xindetai <xindetai@qq.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7.5"
torch = "^1.10"
pytorch-lightning = "^1.5.8"
matplotlib = "^3.5.2"
torchaudio = "^0.11.0"
seaborn = "^0.11.2"
SoundFile = "^0.10.3"
scikit-learn = "^0.22"
numpy = "^1.20"
pandas = "^1.3.4"
hydra-core = "^1.1.2"
transformers = "^4.19.1"
fairseq = "^0.10.2"
speechbrain = "^0.5.11"
webrtcvad = "^2.0.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
